crypto: Entropy Source and DRNG Manager
The kernel crypto API contains deterministic random number generators
(DRNG) which a caller must seed and reseed. The task of seeding a DRNG
is a non-trivial task requiring the consideration of a significant
number of aspects. The Entropy Source and DRNG Manager (ESDM) fills
that gap to transparently seed and reseed DRNGs. A user of the ESDM
obtains random numbers from an appropriately seeded and initialized
DRNG. Further, the ESDM controls various entropy sources guaranteeing
that they are properly initialized and managed.

The ESDM consists of two main parts:

- The entropy source (ES) manager implemented in lrng_es_mgr.c
  controls the available entropy sources including pulling appropritate
  amount of data from them for the DRNG manager.

- The DRNG manager provided with lrng_drng_mgr.c controls the DRNG(s)
  and ensures proper seeding and reseeding.

The entropy source manager controls the entropy sources registered in
the lrng_es array. The entropy sources provide a function pointer data
structure that is used to obtain the services from it.

The ES manager triggers the initial seeding of the DRNGs during boot
time in three stages:

1. The DRNG is seeded from the entropy sources if all entropy sources
   collectively have at least 32 bits of entropy available. The goal
   of this step is to ensure that the DRNG receive some initial entropy
   as early as possible.

2. The DRNG is reseeded from the entropy sources if all entropy sources
   collectively have at least 128 bits of entropy available.

3. The DRNG is reseeded from the entropy sources if all entropy sources
   collectively have at least 256 bits of entropy available.

At the time of the reseeding steps, the DRNG requests as much entropy as
is available in order to skip certain steps and reach the seeding level
of 256 bits. This may imply that one or more of the aforementioned steps
are skipped.

In all listed steps, the DRNG is (re)seeded with a number of random bytes
from the entropy pool that is at most the amount of entropy present in
the entropy pool. This means that for example when the entropy pool
contains 128 or more bits of entropy, the DRNG is seeded with that amount
of entropy as well.

Entropy sources (ES) inform the ES manager when new entropy has been
collected using the lrng_es_add_entropy() function. That function
schedules a DRNG (re)seed with the DRNG manager. When the DRNG manager
requests entropy data, the function lrng_fill_seed_buffer fills the seed
buffer by iterating through all available ES. The output of all entropy
sources is concatenated with each other. Further, the seed buffer
contains the amount of entropy each entropy credits its data. Finally a
time stamp is added.

The ES trigger such (re)seeding events only as long as not all DRNGs
are fully seeded with 256 bits of entropy. Once that seeding level is
reached, the triggers are not further processed.

The DRNG manager initializes the initial DRNG instance during late stage
of the kernel boot process before user space is triggered. The DRNG is
seeded at the following occasions:

- when the DRNG is initialized, the available amount of entropy is used,

- during boot time until the DRNG is fully initialized, the reaching of
  the aforementioned seeding steps of 32/128/256 bits of entropy trigger
  a reseed of the DRNG.

- at runtime after the elapse of 600 seconds since the last seeding,
  the DRNG reseeding flag is raised

- at runtime when more than 2^20 generate operations were performed by
  the given DRNG since last reseeding, the reseeding flag is raised

Raising the reseeding flag implies that the DRNG is seeded in process
context the next time a caller requests random numbers.

At runtime, the DRNG manager requires at least 128 bits of entropy from
the entropy sources (or 256 bits when the FIPS mode is active to be
SP800-90C compliant). It may be possible that the entropy sources may
not deliver that amount. The DRNG is reseeded with the available amount
of entropy and continues to operate. Yet, when after 2^30 generate
requests since the last seeding with 128 bits (or 256 bits in FIPS mode)
the DRNG cannot be seeded with 128 bits (or 256 bits), the DRNG becomes
unseeded which means it will not produce random numbers until it is
fully reseeded again.

To support the DRNG manager, a DRNG implementation is provided with
lrng_drng_kcapi.c. It uses the kernel crypto API RNG framework and
allows the specification of the used DRNG with the kernel command line
option of lrng_drng_kcapi.drng_name. If no reference is given, the
default is the SP800-90A DRBG. In case the chosen DRNG requires the seed
to have a certain length, a hash is used bring the entropy buffer into
the proper size.

In addition, the DRNG manager controls the message digest implementation
offered to entropy sources when they want to perform a conditioning
operation. As entropy sources may require the conditioning operation at
any time, the default is a SHA-256 software hash implementation that
neither sleeps nor does it need any memory allocation operation.
Therefore, this hash is available even for the earliest kernel
operations.

The initial drop of the ESDM includes the entropy source of the
"auxiliary" pool. This entropy source must always be present. It is an
entropy pool that is based on the state of a message digest. Every
insertion of data is a hash update operation. In order to obtain data, a
hash final operation is performed. The purpose of this auxiliary pool is
twofold:

- Provide a general interface to inject an arbitrary amount of data from
  any external source. When providing such data, the caller may specify
  the amount of entropy it contains.

- The auxiliary pool also provides the backtracking resistance for all
  entropy sources. Once a seed buffer is filled from all entropy sources
  it is re-inserted into the auxiliary pool at the same time it is used
  for seeding the DRNG. Naturally, the insertion of the seed buffer into
  the auxiliary pool is not credited with any entropy.

If enabled during compile time with the option of
CONFIG_CRYPTO_ESDM_OVERSAMPLE_ENTROPY_SOURCES, the entropy source
oversampling is activated. However, this oversampling is only enforced
if the kernel is booted with fips=1. The oversampling pulls 128 more
bits of entropy than originally requested. This implies that when 256
bits of entropy are requested for a (re)seed of a DRNG, the ES are
queried for 384 bits. This oversampling complies with SP800-90C.

This patch set contains a number of header files for subsequent
additions, but with the current Kconfig settings, these additional
settings will be folded to noops.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>